Refactor: Extract professor extra information fetching logic into a dedicated fetcher and Riverpod provider.
